+++ packages/query/src/index.ts
@@ -3,6 +3,6 @@ export * from "./types.js";
 export { tokenize, ftsQuery } from "./terms.js";
 export { highlight } from "./highlight.js";
 export { regionOf } from "./moment.js";
-export { ollamaChat } from "./ask.js";
+export { ollamaChat, claudeChat, CLOUD_ASK_MODEL, type ChatFn, type AskDeps } from "./ask.js";
 export { PolicyDeniedError } from "@brainlog/policy";
 export { pulse, sessionize, weekBounds, type PulseResult } from "./pulse.js";